+++ src/main/java/io/datajek/moneytransferrest/service/UserService.java
@@ -1,20 +1,16 @@
 package io.datajek.moneytransferrest.service;
 
-import io.datajek.moneytransferrest.dto.CredentialsDTO;
+
 import io.datajek.moneytransferrest.dto.TransactionDTO;
 import io.datajek.moneytransferrest.model.TransactionEntity;
 import io.datajek.moneytransferrest.model.UserEntity;
-import jakarta.servlet.http.HttpSession;
-import org.springframework.http.ResponseEntity;
-
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface UserService {
     TransactionEntity transferMoney(TransactionDTO transaction, UserEntity sender);
-    UserEntity findById(int id);
+    UserEntity findById(long id);
     List<UserEntity> findAll();
     UserEntity save(UserEntity p);
-    UserEntity update(int id, UserEntity p);
-    void delete(int id);
+    UserEntity update(long id, UserEntity p);
+    void delete(long id);
 }
